Split input into array and clean them.
Take the string input, and convert it into an array of strings since
it can be multiple words. Also clean each individual word so that
they are all lowercase and have punctuation removed (except for
apostrophes). This is so that comparisons will be easier since they will
now match the format of the dictionary being lowercase and not having
punctuation.