Generate all substrings, and update dictionary
Should've split this into 2 commits.

But first, write function all_substrings_of_word in order to get
all the substrings of a word. This makes the comparison easier
since you then just iterate over all of those substrings.

Then for each of those substrings, if it is in the dictionary,
update the output hash.

End result is a hash relating each found substring to the number
of times it was found.